// API/src/enums/logMessage.ts
enum LOG_MESSAGE {
    SERVER_CONNECTED = "Server Connected to Mongoose",
    SERVER_NOT_CONNECTED = "Error Connecting to Mongoose => Error: ",
    MONGOOSE_CONNECTED = "Mongoose Connected to: ",
    MONGOOSE_DISCONNECTED = "Mongoose Diconnected",
    MONGOOSE_ERROR = "Mongoose Connection Error:",
    CONNECTION_CLOSED = "Connection Closed",
    SERVER_STARTED = "🚀 Server Started at: "
}

export default LOG_MESSAGE;